Refractored the typings for database models
