--- typings/user.ts
@@ -1,4 +1,0 @@
-export interface SessionUser {
-  id: string;
-  username: string;
-}
